Check for file existence
My environment failed to create a any rows in asmdnsname and therefore did not create the asmvmlist.json file.  As a result it would fail when trying to delete the file.  

I also added a note about being sure you are running as the Service administrator or the Co-administrator privileges or the list commands using the subscription will not work.
https://stackoverflow.com/questions/36125930/error-fetching-vm-list-using-azure-cli-tools